Started Audio work
Upon startup, the howler library is included. Created a mutation observer that will be listening for tags which represent songs. The songs will have their locations stored inside a file called map.json which will exist in the audio directory. The json will tell the program where to find the audio files on github